Move compression to ztoc as a sub package
