+++ fs/backgroundfetcher/resolver.go
@@ -23,9 +23,9 @@ import (
 	"sync"
 	"time"
 
-	"github.com/awslabs/soci-snapshotter/compression"
 	commonmetrics "github.com/awslabs/soci-snapshotter/fs/metrics/common"
 	sm "github.com/awslabs/soci-snapshotter/fs/span-manager"
+	"github.com/awslabs/soci-snapshotter/ztoc/compression"
 	"github.com/containerd/containerd/log"
 	"github.com/opencontainers/go-digest"
 	"github.com/sirupsen/logrus"
