+++ ztoc/compression/gzip_zinfo.c
