+++ ztoc/compression/gzip_zinfo.go
