+++ ztoc/compression/gzip_zinfo.h
