+++ ztoc/compression/gzip_zinfo_test.go
