+++ ztoc/compression/types.go
