+++ ztoc/compression/zinfo.go
