+++ ztoc/toc_builder_test.go
@@ -22,8 +22,8 @@ import (
 	"os"
 	"testing"
 
-	"github.com/awslabs/soci-snapshotter/compression"
 	"github.com/awslabs/soci-snapshotter/util/testutil"
+	"github.com/awslabs/soci-snapshotter/ztoc/compression"
 	"github.com/klauspost/compress/zstd"
 )
