+++ ztoc/ztoc_test.go
@@ -25,8 +25,8 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/awslabs/soci-snapshotter/compression"
 	"github.com/awslabs/soci-snapshotter/util/testutil"
+	"github.com/awslabs/soci-snapshotter/ztoc/compression"
 	"github.com/opencontainers/go-digest"
 )
